Add a standalone RPC stub generator.
This generator can be called from the usual CAmkES entry point
(`camkes.sh stubgen ...`), but is expected to be used outside the context of a
CAmkES system. The purpose of this is to allow CAmkES IDL to be used in
specifying non-CAmkES systems. The examples given are fairly primitive, but
should be enough to guide template authors in implementing their own RPC stub
code.