Template context: Add a mechanism to write C-style include guards.
Jinja provides two mechanisms for including content from another template file,
`import` and `include`. Import, by default, does not include your current
context and imports the target as a module. The result is that any raw text in
the file is not directly included. Include includes your current context, but
does not let the included file modify the context.

For AutoCorres proofs, it is desirable to abstract some general proofs into
templates that we include with a C-like #include mechanism. However the design
described above makes neither suitable as-is for this purpose. This commit adds
a global set that can be used to track the included files and be modified in
included contexts using `do` statements.

Whereas a C header guard would look like:
 #ifndef _MY_FILE_H_
 #define _MY_FILE_H_
 ...
 #endif
One of these Jinja guards could be written as:
 /*- if 'my_file.h' not in included -*/
 /*- do included.add('my_file.h') -*/
 ...
 /*- endif -*/