Fix: Check for reference types when specialising syscall stubs.
The code at these locations attempts to determine whether the arguments to an
RPC call can be packed into registers and hit the kernel IPC fastpath.
Unfortunately it assumed that all types of your methods were CAmkES native
types, which was not necessarily true. This commit checks whether we have any
reference types (C typedefs) in the method we're currently looking at
specialising.

Closes JIRA CAMKES-303