Fix: Reset badge to zero before polling in AsynchNative connector.
Prior to this, we could incorrectly conclude we'd received an event when we
hadn't as the badge had been set from the previous time we'd called poll. Note
that this behaviour was actually masked by a larger issue in the kernel API
(SELFOUR-224). The fix for the kernel issue actually masked the behaviour here
in a different way, so this was still not exactly a problem, but it wasn't very
robust against future API implementation changes. The kernel changes also
resolved the second part of the CAMKES JIRA issue.

JIRA CAMKES-311