Optimisation: pre-compile templates when the compilation cache is enabled.
This commit makes some changes that kick in when the compilation cache is
enabled. If it is the first time we have seen the templates, they are compiled
into Python byte code and stored in the compilation cache. When rendering
templates after this, we can just retrieve the pre-compiled version from the
compilation cache.

It is difficult to measure the effect of this because it is optimising for a
case when you have edited your ADL (cache miss on the rendered template itself)
but have not modified CAmkES itself. A contrived benchmark gives an optimistic
performance improvement of 5%, but too much weight should not be given to this
figure.