Fix: Remove 'development' changes that were accidentally transplanted.
This commit removes a relocation of the IPC buffer and thread stacks to a
separate ELF section. This relocation only makes sense in the context of a
dynamic linker script that is not present on the 'master' branch. Whether these
changes should end up on 'master' eventually is an open question, but for now
we just remove this malformed change.