Optimisation: Don't deepcopy the AST when passing it to filters/templates.
The original rationale behind this was that we might be invoking defective
templates that illegally modify the AST during their execution. If this is the
case, then I think it's reasonable to assume we've lost already, considering
the templates are essentially gifted the power of code injection.

Note that this is a change that has been present in 'development' for some
time.