Optimisation: Don't deepcopy AST entities when passing them to templates.
As of 1e50795a727c4a473ab31a0c452e8ca9a515585e (and prior, really) we're
trusting user templates and filtering code^1. This commit goes further along
this line, passing them a modifiable version of the current AST entity, though
note that they are expected not to modify it. Empirically, this shaves ~25
seconds off a non-cached build of the VM project.

1. Realistically, it never made sense not to trust these pieces of code as they
   can cause all sorts of chaos anyway, but these copies guarded against buggy,
   rather than malicious code.