Implement 'refin' as a new parameter direction.
This commit implements a new direction for RPC parameters, 'refin.' In
everything except the C backend, this direction is identical to 'in.' In the C
backend this generates a user function prototype that passes this parameter by
pointer instead of by value. The primary use case of this is for interfaces
involving C types that are typedefed structs. The cost of a function call
involving such types can be decreased by passing the parameter by pointer. In
generaly this should only be used as an optimisation and 'in' should remain the
default direction for input parameters.