seL4RPCCall: Only elide asm clobbers when explicitly trusting components.
This commit introduces an extra degree of safety, wherein users are required to
explicitly use an attribute to declare component instances 'trusted = "true"'
if they want to elide all clobbers in the specialised asm stub. If the partner
you are communicating with is not trusted it is not safe to elide these
clobbers as they can overflow the message length you wre expecting and cause
unexpected behaviour. To my knowledge no one had hit issues like this or were
using this specialisation in the presence of untrusted components, but this is
an added safety measure.

In future it would be possible to introduce more broader stub optimisations to
use seL4_CallWithMRs when we can detect (at generation time) that there is a
payload, but it is short enough to fit in the message registers.