Optimisation: alloc destination space for strings on-demand.
This commit removes the constant `CAMKES_STR_MAX` and references to this that
were used as arguments to `malloc`. Instead we now allocate through `strdup`
on-demand as we always know, in advance, the length of the string we're
unmarshalling. This change is an optimisation for memory usage and not
necessary for correctness, though it does have user-visible effects. Users
should be aware that string arrays are now a malloced array of pointers to
other malloced buffers, as opposed to a single malloced (2D) array.