Optimisation: Cache virtual-to-physical address translations.
Doing DMA on a CAmkES system requires specifying a DMA pool, which results in a
region of your address space backed by frames that are "reversible" (you can
ask the kernel what their physical address is). This commit caches these
reverse lookups in static variables. The result is that we avoid the syscall
overhead when repeatedly reversing DMA frames, as can happen in a naïve DMA
driver.