Support for automatic promotion of large frames.
In large CAmkES systems, compilation time and load time can be dominated by
creating and managing numerous 4K frames. We can speed up both processes by
backing a component's address space with larger frames as supported by the
platform. This commit transplants support for this from the development branch
of CAmkES. Large frame promotion is enabled by default, but can be selectively
disabled in the build system.

This commit also contains preliminary support for running in HYP mode on ARM.
Support for this is not yet in the kernel, but these changes should provide a
relatively smooth transition when this does arrive. No changes towards
supporting PAE are included in this commit.

Closes JIRA CAMKES-145 (requires further work on development; see JIRA
SELFOUR-257).