Fix: Remove orphaned dataport frames.
The CapDL filters are responsible for matching up two sides of a dataport such
that each component's mapping points at the same physical frame(s). This
functionality previously updated the mapping with no regard for the previously
hosted frame. As a result, we were producing CapDL specs with orphaned
(unreachable) frames which wasted memory and runtime. This commit fixes this
issue.

Closes JIRA CAMKES-345