libsel4camkes: Allow page-colocated DMA allocations.
The previous allocation scheme was overly conservative and rounded all DMA
allocations up to page boundaries as a security precaution. This extra step
actually serves no purpose in a system without an IOMMU, where a misconfigured
DMA request is going to access incorrect memory physically anyway. The effect
of this commit should be permitting more DMA allocations with a smaller pool.