Use ROUND_UP_UNSAFE in glue code instead of ROUND_UP.
ROUND_UP_UNSAFE is an alternative to ROUND_UP that evaluates its arguments more
than once and thus is sometimes unwise to call. In the context we are invoking
it, it is safe to evaluate the arguments multiple times. The motivation for
this switch is because the verification C parser cannot handle the GNU
statement expressions used in ROUND_UP.