seL4RPC/seL4RPCCall: avoid calls to `memset` where possible.
We were using `memset` to construct a dummy return value when failing glue code
functions. This involved taking the address of a local, which is not compatible
with the verification C parser. We now return a literal when using a return
type that is a native CAmkES type. External C typedefs still require a
`memset`.

This change has no effect on binary code as GCC can - and does - freely convert
assignments and `memset`s back and forth.