Marshalling: create method indices as static consts.
The primary motivation for this change is to avoid taking the address of local
variables, which upsets the verification C parser. Surprisingly it also has a
reduction on the size of generated code. This may not necessarily be
significant as the cost of marshalling other arguments will generally subsume
the method index itself.