Marshalling RPC inputs: unwind pointer dereferencing.
Prior to this commit, the glue code eagerly dereferenced pointers and passed
input parameters by value to their specialised marshalling functions. This
caused a requirement to take the address of a local variable in the specialised
marshalling functions. Taking the address of a local variable is outside the
scope of verification and we need to avoid this. This commit now carries the
pointer, where possible, right through to the `memcpy` invocation.

It is still necessary to take the address of a local backing an `in` parameter.
This will need to be avoided in future with TLS. There are further changes
along these lines that need to be done for the other marshalling/unmarshalling
code. Once again, this has no effect on compiler-generated code in observed
cases.