Support for configurable TLS.
For verification we need to avoid taking the address of local variables. In the
glue code this is difficult because multiple threads may be executing in the
glue concurrently. The current work around for this is per-thread global
variables that can have their address taken. This is added here as a
configuration option with the default remaining the existing behaviour, because
this structure has a non-trivial performance impact. In particular, GCC fails
to determine that our use of static variables does not rely on their previous
value and thus persistent stores to them can be eliminated. Experimenting with
small examples suggests that this is just not an optimisation that GCC ever
attempts. This is unsurprising as this pattern of variable usage is fairly
unusual and would be unlikely to be encountered in handwritten code.

This change should have no visible effect to unaware users.