libsel4camkes: Make the DMA allocator agnostic to page size.
This commit adjusts the API for the DMA allocator to allow the caller to
specify the page size used in the mappings backing the DMA pool. It also allows
the caller to instruct the allocator itself to automatically determine the page
size. The purpose of this is to future proof this code against potential
changes to allow the DMA pool to be backed with large frames. No users are
expected to be manually calling the DMA initialisation function, so this is not
expected to have any user-visible effects.