Tweak how generated files are pruned.
To be more amenable to verification, we need to make some slight adjustments to
guide the pruning process to both exclude things that we can't handle and
retain or force-emit things we need. None of this should compromise the
correspondence between verification and the running system as, when we build in
this configuration, we compile the same pruned file that verification sees.