Cherry-pick initial AutoCorres backend from 'development'.
This commit adds the infrastructure for the AutoCorres backend to CAmkES, as
well as the previous version of theory generation. This covers no-failure
proofs for various glue code. Note that this commit also includes some general
refactoring of the seL4RPCSimple connector.

The infrastructure, config and targets for the unified RPC theory are also
included in this commit, though these are stubbed out currently. In future, it
should be trivial to transfer this across as well, but I wanted to stabilise
and clean things up in development a little first.