Cherry-pick unified theory AutoCorres templates from 'development'.
This commit adds the unified theory part of the AutoCorres backend that has
been present in development for some time. It is fairly experimental and there
are some known bugs for 'inout' parameters currently. To work on it, you will
need an installed and configured copy of l4.verified and a moderately deep
understanding of AutoCorres and the Isabelle simplifier. There's virtually no
documentation on this backend or its templates, but if you want to use it
please ask me questions.