seL4RPC/seL4RPCCall: Rename internal glue code functions to avoid collisions.
This commit renames marshalling and unmarshalling functions that previously had
the same name. These functions were compiled into separate binaries, so this
was not an issue from an implementation perspective, but this caused problems
for verification, where we see both snippets of generated code as the same
translation unit. This commit is expected to have no user-visible effects and
no effect on generated binaries.