Suffix TLS variables with the connection side of their host.
The purpose of this commit is to avoid naming conflicts with the TLS variables
from opposing sides of a connection and their accessors. As for the previous
commit, this is irrelevant for the implementation, but aids verification.
Again, this is expected to have no user-visible effects and no effect on the
generated binaries.