Qualify input pointers as `const`.
** BREAKING CHANGE **
  Functions that correspond to RPC interfaces may need `const` added to some of
  their arguments.

This commit alters the prototypes for RPC interfaces to mark input pointers --
specifically 'refin' scalars, 'in' strings and 'in' arrays -- as `const`. The
purpose of this is largely a hint to user programmers, but it can potentially
enable some further compiler optimisations. An effort has been made to maintain
const-ness of pointers internally for the purposes of marshalling as well.

Note that `const` has not been applied to any parameter that manifests as a
double pointer ('in' string arrays, 'refin' strings, 'refin' arrays). The
rationale for this decision was that introducing `const` here actually makes
things *more* complicated for the user. E.g. a `char**` does not naturally
coerce to a `const char**`. The result is a series of confusing compilation
warnings and generated code that the compiler (incorrectly) believes to be
unsafe.

Closes JIRA CAMKES-358