Fix: Mark AST objects as referencable _after_ they have been fully digested.
This commit fixes a subtle issue whereby it was possible to reference your
parent (or grandparent, etc.) in a CAmkES specification. Beyond wasting runtime
by considering these invalid referents, in the wake of support for hierarchical
components this likely allowed you to create cycles in the AST. No current
applications are known to actually depend on this behaviour, so this change
should be largely transparent to users.