Wrap dataport array typedefs in structs to reduce surprises in user code.
Both the `Buf` typedef and the internal area defined for a dataport were byte
arrays. When a user was not consciously aware of these being arrays, it was
possible for them perform casts and pointer operations that had surprising
outcomes. This commit makes both of these structs, and then referenced by
pointers. The in-memory data layout should be unaffected, but this should
reduce surprise misbehaviours in user code.

Closes JIRA CAMKES-369