Check at compile-time that RPC parameters typedefs are not arrays.
The types of RPC parameters can be external references that are expected to be
provided as visible C typedefs at compile-time. In C it is possible for a user
to provide a typedef representing an array type. Such types have different
behaviour to scalar typedefs, in particular the way they are transformed across
function calls.

Typedef-ed arrays like this are not supported as RPC parameters, due to the
glue code's limited ability to introspect these types. This commit adds a best
effort mechanism to induce a compile-time warning if the user has made use of
such an unsupported type.

Closes JIRA CAMKES-306