Fix parsing of numbers to correctly identify hex values
The previous regex would select on the first of the two ORd regex patterns that matched
and then pick the longest token from that. This means that 0x2E would match on the part
that had lowercase characters, leaving a trailing E in the token stream. Allowing mixing
of upper and lower case hex literals is the easiest way to fix this