Optimisation: Move second compilation cache check earlier.
The second time we check the compilation cache is intended to serve requests
for items when the input files have changed, but in a semantically meaningless
way (e.g. modifying a comment or whitespace). A cache hit allows us to bail out
of the runner early saving compilation time.

This check was being performed later than it needed to be. In particular the
following operations were occurring unconditionally, none of which the check
depended upon:
 * Backup of the original AST;
 * Deduping the AST;
 * Reference resolution;
 * Reference collapse;
 * Miscellaneous AST transformations; and
 * Hierarchical component resolution.

This should have an appreciable effect on compile times.

Conflicts:
	camkes/runner/__main__.py