Remove --set command line argument.
This command line argument was previously introduced for a verification use
case of passing auxiliary information in to the template context from the
environment. No one is currently using it and it is not especially clean. In
particular, it (intentionally) allows command-line code injection into the code
generator.

You can still achieve extension of the template context with passive data using
environment variables.

Closes JIRA CAMKES-361