Fix: guard shared variable registration with CSpace, instead of entity.
It was fine to guard shared variable registration with either for the
per-component template, but in connection templates the "entity" in the
connection itself. The intention was actually to guard this registration with
something that corresponded to the address space of the local name. Using the
CSpace isn't perfect, but it's the best we can do for now.

JIRA CAMKES-323