Fix: cope with PYTHONPATH being unset in the user's environment.
`PYTHONPATH` is an environment variable used by Python to search for imports,
in a similar way to that in which `PATH` is used by your shell. The standard
idiom for appending a path to your `PATH` variable is something like:

    PATH := ${PATH}:foo

This is fine for `PATH` because it can reasonably be assumed that the user's
environment has at least one path already in `PATH`. Unfortunately this
assumption is not true for `PYTHONPATH` and many users have this unset. As a
result, the same idiom accidentally adds an empty string to the list of search
paths, which equates to the current directory. The effect of this is that the
current directory is searched for Python includes _in preference_ to
directories relative to the source of the `import` statement.

This effect was probably not observable by users, but caused problems during
development that were difficult to diagnose. Following this commit, unset or
empty `PYTHONPATH` variables should not cause problems.