Fix: Only ever remove CapDL objects through the object space, never directly.
This code was directly accessing the object collection of the object allocator
in order to remove objects. While no effects of this were visible at runtime,
this induced a stale label mapping. Ultimately this results in an inconsistent
information flow view of the system. Not critical, as this is detected in the
follow on proofs, but this made information flow reasoning impossible.

Partial fix for JIRA CAMKES-389