Extend label mapping with translation from CapDL symbol to Isabelle symbol.
This mapping is a convenience for one of the CAmkES verification passes. We'd
like to be able to reference objects by ID, but their CapDL ID is a little
awkward to compute. The CapDL translator gives us Isabelle symbols that map to
these IDs, but it is not straightforward to translate from the string
representation we have to one of these symbols. Thankfully we can do this
trivially at theory generation-time, as is done in this commit.