remove -u flag to cp for posix compatibility
cp -u is a GNU coreutils extension that's not present in the BSD
versions of the tools. There's no portable equivalent to `cp -u` that I
can find (without introducing a dependency on rsync), so just drop the
flag and copy all files every time. In my tests, this has a negligible
impact upon build time.