Hardware: include stdint.h automatically when using IO ports.
This has very little impact currently and should have no user-visible effects.
The purpose of this commit is support for upcoming changes that remove
[u]int*_t as built in types from CAmkES. At this point they become opaque C
types that the user is expected to have in scope when using. In the case of the
integer types, this is usually not necessary because CAmkES coincidentally
includes these as part of other operations, but for correctness we should do it
explicitly.