Remove support for arrays as return types.
This commit effectively backs out part of
20f4072ac25eeb707f4e17c4a26501d4b59308ee. The motivation for removing support
for this is that array return types are not representable in the upcoming
refactored AST. Note that there is currently no grammar for describing an array
return type anyway, so no end users are relying on this support. The only way
to use an array return type was to programmatically construct a method with one
and I am not currently aware of any uses of this.