Tweak external objdump options to default to auto-discover.
The CAmkES code generator has an option to call out to your toolchain's objdump
to retrieve ELF symbol information instead of using pyelftools. The advantage
of this is that it turns out to be significantly faster. Traditionally this was
disabled as it was seen as more fragile. There have been no significant
difficulties with this functionality, so this commit alters the default
behaviour to auto-detect when objdump is in your PATH and use it. You can still
manually set this feature off or on if necessary.