Remove --post-render-edit option.
This option was intended to allow developers to drop into an editor after the
code generator has run in order to manually tweak templates. In practice, this
was useful for a while, but has now been more than superseded by the ability to
drop into the Python debugger from the template context.