Optimisation: Disable auto-reloading of templates.
This (default) feature of Jinja is designed for an environment like a
webserver, where templates are expected to change on disk while the service is
still running. CAmkES' code generator does not work as a persistent service and
templates are not expected to change while the generator is running. The net
effect of this should be a (negligible) performance improvement.