Remove usage of implicit `tls` variable in verification templates.
This variable was expected to be set from the command line, though the
mechanism by which this was accomplished was removed in
213cc215a2525a9bba8192a159f223095a7c1367. It's not clear there was an advantage
to be had by suppressing this definition anyway. While lemmas cost us a lot in
terms of runtime, definitions are comparatively cheap. We now simply always
emit the TLS-relevant pieces.