Squash some compiler warnings resulting from the parameter array type check.
In certain configurations your C compiler would warn about passing an
uninitialised value in a function call, though the value of the variable in
question was never referred to. I suspect this was due to the compiler's
inability to look inside the eventual `asm` block, but I did not investigate
this. We work around it in a fairly questionable way by just declaring the
variable `static`. Note that simply initialising this variable would have been
non-trivial as we would have needed to deduce a valid literal of its type.