Throw generation-time exceptions when dealing with unsupported types.
This older verification template was using a textual marker to indicate
unsupported types. The result was that, instead of a generation-time failure,
your eventual proof would fail with a syntax error. This was a little unclean
and also led you to discover the failure much later in your workflow. This
commit cleans it up into a generation-time exception like other such cases.