Use `mkdirp` in preference to `os.makedirs`.
When running multiple instances of the CAmkES code generator, a race condition
is sometimes provoked whereby a call to `os.makedirs`, after checking the
target directory does not exist, can fail because another instance of the code
generator created the directory in the meantime. This change makes the code
generator robust against such effects, allowing multiple instances of it to
execute in parallel.