Fix: Rename control TCB to avoid symbol collisions.
**BREAKING CHANGE**

This commit renames the control TCB from "tcb__control" to "tcb_0_control". The
purpose is to avoid unintentional symbol collisions with TCBs belonging to
interfaces named "_control". As part of the same disambiguation process, any
attributes that were of the form "_control_<property>" are now "_<property>".
Because an interface cannot have the empty string as a name, this avoids
collisions.

This is marked as a breaking change because users who adjust attributes of the
control thread (priority, stack size, domain, etc.) will need to change the
attribute they are setting. It is possible to set both the new and the old
attribute in order to smooth the transition across this changeset.