Pad DMA frame indices.
This commit changes the CapDL names of frame objects that back the DMA pool to
pad their indices to four digits. Though seemingly irrelevant, this commit has
an important practical benefit. The CapDL translator alphabetically orders
objects of the same size when translating a specification, and this order then
goes on to determine where objects are allocated in physical memory. By
ensuring that the alphabetic order of the DMA frame objects matches the order
of their mapping addresses in virtual memory, we increase the chances of their
being contiguous in physical memory. Thus we increase the possibility of
satisfying large DMA allocations that require both physical and virtual
contiguity.

This explanation sounds pretty contrived and far fetched, but in current
systems this is the difference between a maximum of 40KB of usable DMA memory
vs 4MB of usable DMA memory.