Use central name mangling support for fault handler detection.
Prior to this commit, one of the filters hard coded the name mangling algorithm
for fault handlers in order to detect the TCB of a fault handler. This was
unnecessary as the central name mangling support is accessible from this
context. This commit has no effect on functionality, but simplifies future
maintenance by obviating the need to modify the filter when the name mangling
scheme is changed.