Use a collision-resistant TCB name mangling algorithm.
Prior to this commit, it was possible to generate a naming collision in a CapDL
spec using a specific input spec. For example, in the following:

    component Foo {
      uses Proc tcb;
    }

    component Bar {
      provides Proc tcb_tcb;
    }

    assembly {
      composition {
        component Foo abc_tcb;
        component Bar abc;

        connection seL4RPC c(from abc_tcb.tcb, to abc.tcb_tcb);
      }
    }

two CapDL objects named "tcb_abc_tcb_tcb_0000" would be generated. This was
symptomatic of a TCB name mangling algorithm that permitted collisions. This
commit moves to an algorithm designed to avoid such collisions. The above
example would now generate TCBs "abc_tcb_7_tcb_3_0000_tcb" and
"abc_3_tcb_tcb_7_0000_tcb".

As a more broad description of the algorithm, interface TCBs are named:

    <instance>_<len(instance)>_<interface>_<len(interface)>_<intra_index>_tcb

Fault handler TCBs are named:

    <instance>_<len(instance)>_0_fault_handler_15_0000_tcb

Note that this still fits in with the previous coincidental naming in that
these TCBs can still be treated as interface TCBs in certain contexts. Control
TCBs are named:

    <instance>_<len(instance)>_0_control_9_tcb

The above scheme is intended to be resistant to all possible collisions. Note
that this commit also moves the "tcb" tag to the end of the name, which ensures
that these objects can also no longer collide with endpoints or CNodes.

Closes CAMKES-418