Always pass a cap to an AEP when associating IRQs.
Prior to this change, templates that associated IRQs with AEPs always did so via
an AEP object. Internally, the python-capdl module would allocate a cap to this
object, as IRQs always need to be associated with AEP *caps*, as opposed to AEP
*objects*. This change modifies all in-tree instances to pass a cap to the AEP.
This is marginally more efficient (due to avoiding creating another cap in
python-capdl, but the main motivation for this change is that an upcoming
change to python-capdl removes the ability to pass an AEP object as this
parameter and requires a cap.