Added "hardware_cached" attribute to hardware dataports
Setting the <instance>.<interface>_hardware_cached to the boolean True
enables caching for that region of memory. Setting it to False
disables caching. The default value is False.

Frames backing hardware dataports are allocated by a template

Previously they were allocated in the collapse_shared_frames filter.
Moving their allocation to the template is necessary to support
components being able to flush hardware dataport frames from the cpu
cache. This is necessary because the cap addresses of the frames must be
included in the hardware dataport template. In the
collapse_shared_frames filter, the frames backing hardware dataports are
located by searching the cspace of the component connected to the
dataport.