Use --architecture value in preference to --hyp for determining Hyp mode.
This commit alters the runner to use the value of the --architecture command
line argument to decide whether we are targeting ARM's Hyp mode or not. This is
in preference to the previous practice of using the --hyp command line
argument. The eventual goal is to deprecate and remove --hyp, in order to unify
the handling of Hyp mode.

Related to JIRA CAMKES-334, JIRA SELFOUR-297