Use --architecture value instead of ELF architecture for paging calculations.
This commit modifies some of the code in Filters.py to use the architecture
indicated to the runner on the command line rather than that of the subject ELF
file to make decisions about paging structures. This is only possible now
because python-capdl recognises the architecture string "arm_hyp". This paves
the way for removing any specific handling of Hyp mode.

Note that this commit also has the accidental effect of making
`collapse_shared_frames` Hyp-aware. This function does nothing special to
handle ARM's Hyp mode which, as far as I can see, means it is broken on a Hyp
kernel. With that in mind, this commit does nothing to break it any further,
though it also does not fix it.

Related to JIRA CAMKES-394, CAMKES-334, SELFOUR-297