Only put pre-processor headers on copied C-like sources.
This commit modifies a Makefile line that introduces CPP line directives
indicating the original sources of your compiler inputs. The intent of the
original line was to only stamp C/C++ sources this way. However, though CAmkES
itself does not deliberately produce other files that end up passing through
this rule, it was possible for the user to (intentionally or accidentally) dump
other junk in their build directories which was then picked up and stamped. In
particular, one affected user had an ELF object file prefixed with this leader
which invalidated it.

This commit guards against this, admittedly obscure, case. Note that this
cherry picks
https://github.inside.nicta.com.au/mfernandez/project/commit/669ad0804307729c234ad11aee092fd4f539197e

Fixes JIRA CAMKES-416