Fix: account for C header includes when stringising a procedure.
This commit fixes an issue where the compilation cache did not notice when you
changed the C header includes in a CAmkES procedure. It is relatively rare for
users to trigger this bug, but it has happened in the past. The underlying
cause is that the stringising functionality (`__repr__`) is overloaded to be
the inputs to the compilation cache as well. This issue is remedied in a more
principled way in the 'next' branch, but for now we correct this imprecision
here leaving the cache architecture intact.

Fixes JIRA CAMKES-380.