seL4HardwareInterrupt: Fix missing check that IRQ number is actually set.
Prior to this commit, it was possible to either deliberately or accidentally
not assign an IRQ number when using the seL4HardwareInterrupt connector. The
result would be silent failure from CAmkES and production of a malformed CapDL
specification. This commit adds a check for this to the template code, making
it more obvious to the user what they need to correct.

Fixes JIRA CAMKES-414