Locate precise line numbers even when PlyPlus throws multiple errors.
Background: When an error is encountered during parsing, CAmkES raises a
`ParseError`. This contains source line number information that is
automatically narrowed when the error is presented to the user. In the case
when the error occurred during PlyPlus lexing or parsing, CAmkES uses a
combination of the PlyPlus exception and C pre-processor line directives to
precisely locate the offending token(s).

Prior to this commit, CAmkES only anticipated PlyPlus exceptions that
identified a single source location. This meant that exceptions that referenced
multiple lines, such as:

    Syntax error in input at 'notifier' (type ID) line 55 col 16
    Syntax error in input at 'client' (type ID) line 60 col 20
    Syntax error in input at 'client' (type ID) line 65 col 20
    Syntax error in input at 'client1' (type ID) line 70 col 19
    ...

caused the line narrowing algorithm to exit early and the user was presented
with an imprecise source location. This commit corrects this imprecision and
such scenarios now give an accurate narrowed line number.